1537 bulk lp tag processing (#1554)
* bulk_create list presence to tags

* Combine tagforms into one save form

* Rework form to handle all chems and tags at once.

* Rework JS and Doc page html to send direct post requests

* JS rework.  Chemicals POSTed as array

* Add env variable for DATA_UPLOAD_MAX_NUMBER_FIELDS
More information on this variable can be found on the [django docs](https://docs.djangoproject.com/en/2.2/ref/settings/#data-upload-max-number-fields).  This variable is a middleware that helps prevent DoS attacks.  In our current case it seems unlikely that is a risk.

* Remove small bugs

* Fix test test_listpresence_tags_form.
This test only appeared passing.  It was saving all 4 associated tags to one ExtractedListPresence.  The test asserted that there were 2 * 2 (ELP count) new connections which just happens to work out as 4 new records.

* Clean up errors on test_hp_cards

* black reformat

* Add some waits to selenium tests

* Remove errant single quote from success message

* Add test_lm_chemical_update staleness check to verify page refresh