2018: fix upload existing reported funcuse (#2024)
#2018   fix upload existing reported funcuse

* alter test to cover capitalization differences

The duplicate-checking logic around lines 610-650 of `/forms/data_goup.py` does not handle capitalization correctly. Each incoming `report_funcuse` is converted to lowercase before being compared to values in the database, so an incoming value with capital letters is not being correctly flagged as a pre-existing string. The test (still failing) now covers this condition.

* Find matching fu regardless of case

Co-authored-by: Matt Frost <mwfrost@gmail.com>